Anton/quantization move (#292)
* Moved quantization.hpp to /common

* Moved quantization.hpp to /common
