+++ test/blas_test.hpp
@@ -43,8 +43,8 @@
 #include <common/float_comparison.hpp>
 #include <common/print_queue_information.hpp>
 #include <common/system_reference_blas.hpp>
+#include <common/quantization.hpp>
 
-#include <utils/quantization.hpp>
 #include "blas_test_macros.hpp"
 
 struct Args {
